Move xterm-internal types back into Types.ts files
