+++ src/Buffer.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { assert } from 'chai';
-import { ITerminal } from '../typings/xterm-internal';
+import { ITerminal } from './Types';
 import { Buffer } from './Buffer';
 import { CircularList } from './utils/CircularList';
 import { MockTerminal } from './utils/TestUtils.test';
