+++ src/Buffer.ts
@@ -3,8 +3,8 @@
  * @license MIT
  */
 
-import { ITerminal, IBuffer, LineData, CharData } from '../typings/xterm-internal';
 import { CircularList } from './utils/CircularList';
+import { LineData, CharData, ITerminal, IBuffer } from './Types';
 
 export const CHAR_DATA_ATTR_INDEX = 0;
 export const CHAR_DATA_CHAR_INDEX = 1;
