+++ src/InputHandler.ts
@@ -4,8 +4,7 @@
  * @license MIT
  */
 
-import { CharData, ITerminal } from '../typings/xterm-internal';
-import { IInputHandler, IInputHandlingTerminal } from './Types';
+import { CharData, IInputHandler, IInputHandlingTerminal, ITerminal } from './Types';
 import { C0 } from './EscapeSequences';
 import { DEFAULT_CHARSET } from './Charsets';
 import { CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX } from './Buffer';
