+++ src/renderer/ColorManager.ts
@@ -3,7 +3,8 @@
  * @license MIT
  */
 
-import { IColorSet, IColorManager, ITheme } from '../../typings/xterm-internal';
+import { IColorSet, IColorManager } from './Types';
+import { ITheme } from 'xterm';
 
 const DEFAULT_FOREGROUND = '#ffffff';
 const DEFAULT_BACKGROUND = '#000000';
