+++ src/renderer/Types.ts
@@ -3,8 +3,8 @@
  * @license MIT
  */
 
-import { IColorSet, IRenderDimensions, ITerminal } from '../../typings/xterm-internal';
-import { IEventEmitter } from 'xterm';
+import { ITerminal } from '../Types';
+import { IEventEmitter, ITheme } from 'xterm';
 
 /**
  * Flags used to render terminal text properly.
@@ -16,6 +16,51 @@ export enum FLAGS {
   INVERSE = 8,
   INVISIBLE = 16,
   DIM = 32
+}
+
+export interface IRenderer extends IEventEmitter {
+  dimensions: IRenderDimensions;
+  colorManager: IColorManager;
+
+  setTheme(theme: ITheme): IColorSet;
+  onWindowResize(devicePixelRatio: number): void;
+  onResize(cols: number, rows: number, didCharSizeChange: boolean): void;
+  onCharSizeChanged(): void;
+  onBlur(): void;
+  onFocus(): void;
+  onSelectionChanged(start: [number, number], end: [number, number]): void;
+  onCursorMove(): void;
+  onOptionsChanged(): void;
+  clear(): void;
+  queueRefresh(start: number, end: number): void;
+}
+
+export interface IColorManager {
+  colors: IColorSet;
+}
+
+export interface IColorSet {
+  foreground: string;
+  background: string;
+  cursor: string;
+  cursorAccent: string;
+  selection: string;
+  ansi: string[];
+}
+
+export interface IRenderDimensions {
+  scaledCharWidth: number;
+  scaledCharHeight: number;
+  scaledCellWidth: number;
+  scaledCellHeight: number;
+  scaledCharLeft: number;
+  scaledCharTop: number;
+  scaledCanvasWidth: number;
+  scaledCanvasHeight: number;
+  canvasWidth: number;
+  canvasHeight: number;
+  actualCellWidth: number;
+  actualCellHeight: number;
 }
 
 export interface IRenderLayer {
